Added Tab for BinSync Utilities (#130)
- added new tab for Utilities
- refactored some GUI code for better naming convention
- added endpoints for Magic Sync and Sync All
