+++ .github/workflows/python-package-gui.yml
@@ -22,7 +22,7 @@ jobs:
         python -m pip install --upgrade pip
         pip install flake8 pytest
         pip install angr-management
-        rm -rf /opt/hostedtoolcache/Python/3.8.12/x64/lib/python3.8/site-packages/angrmanagement/plugins/binsync
+        rm -rf /opt/hostedtoolcache/Python/3.8.12/x64/lib/python3.8/site-packages/angrmanagement/plugins/angr_binsync
         ./install.sh --angr /opt/hostedtoolcache/Python/3.8.12/x64/lib/python3.8/site-packages/angrmanagement/plugins/ --link
 
     - name: Pytest
