+++ binsync/common/ui/panel_tabs/globals_table.py
